# Reduce runtime of go Encode() implementation by 20%