@willbeason 2023-11-09
Playing around with the code, I found a second optimization that improves performance an additional 10%. I'll include it as a follow-up to this PR as it could potentially be controversial.